requirements.txt
Flask
pybadges
gunicorn
requests
python-dotenv